+++ src/core/balance/contract-conditions.ts
@@ -38,17 +38,6 @@ const sortedConditions = computed(() => {
 });
 
 function calculateDeadline(contract: PrunApi.Contract, condition: PrunApi.ContractCondition) {
-  if (condition.type === 'COMEX_PURCHASE_PICKUP') {
-    // The COMEX_PURCHASE_PICKUP condition has unique handling:
-    // Once all its dependencies are fulfilled,
-    // the player needs to pick up the materials using this condition.
-    // For determining the deadline of the COMEX_PURCHASE_PICKUP condition,
-    // we will use the latest deadline among its dependencies.
-    // This is because the materials can be picked up at any time,
-    // making the COMEX_PURCHASE_PICKUP's own deadline irrelevant.
-    return getLatestDependencyDeadline(contract, condition);
-  }
-
   if (condition.deadline) {
     return condition.deadline.timestamp;
   }