+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- `adm-neutral-withdraw-button`: Applies the "neutral" style to the "Withdraw" vote button
 - `fxpo-better-current-price`: Sets the current price to the order book price
 - `fxpo-maximum-lot-size`: Adds a "Maximum Lot Size" field to fill the top order in the order book