+++ CHANGELOG.md
@@ -15,7 +15,8 @@
 - `XIT BURN`: The "Days" column now displays fractional days when less than 10 days remain
 - `XIT BURN`: The "Need" column is now based on fractional days instead of whole days
 - `XIT BURN`: The CSV export now includes the "Need" column
-- Screens with many added elements now render faster
+- `XIT CXTS`: The command now supports an optional "Page Size" parameter
+- `XIT FXTS`: The command now supports an optional "Page Size" parameter
 
 ### Fixed